requirements.txt
streamlit==1.33.0
sentence-transformers==2.6.1
faiss-cpu==1.7.4
requests==2.31.0
numpy==1.26.4
pandas==2.1.3
